+++ changelog.md
@@ -11,6 +11,7 @@ compromise uses semver, and pushes to npm frequently
 - add `#Multiple` tag, and changes to how invalid numbers like 'sixty fifteen hundred' are understood
 - better em-dash/en-dash support
 - better conjugate implicit verbs inside contractions - "i'm", "we've"
+- nouns().articles() method
 <!-- * include 'the #TitleCase' matches in .topics() -->
 
 ##### 11.0.0